% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_coefs_fct.r
\name{best_coefs}
\alias{best_coefs}
\title{best_coefs function}
\usage{
best_coefs(empty_file, leafArea_cm2, delta_max, skip_first, min_CO2,
  max_degree, diagnostic_plots)
}
\arguments{
\item{diagnostic_plots}{}
}
\value{

}
\description{
best_coefs function
}
