% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_list_fct.r
\name{build_list}
\alias{build_list}
\title{build_list function}
\usage{
build_list(path_to_licor_files = "data/",
  pattern_empty = "(mpty).*\\\\.xls",
  pattern_rapidACi = "(fast).*\\\\.xls",
  pattern_standardACi = "(slow).*\\\\.xls",
  pattern_dark = "(dark).*\\\\.xls", timestamp_column = "BN")
}
\arguments{
\item{path_to_licor_files}{Path where all files are stored}

\item{pattern_empty}{Regex pattern that will only match to filenames for empty chamber
files}

\item{pattern_dark}{Regex pattern that will only match to filenames for dark chamber
files}

\item{timestamp_column}{Column index corresponding to the MATCH TIME column in all
Li-Cor Excel files. By default this corresponds to column BN (index = 66) in Excel
files.}

\item{pattern_rapid}{Regex pattern that will only match to filenames for rapid A-Ci
measurement files}

\item{pattern_slow}{Regex pattern that will only match to filenames for standard A-Ci
measurement files}
}
\value{
The function return a dataframe that includes the path to files, the type of
  file and a matching timestamp to an empty chamber file.
}
\description{
Generates a list of the files present in a specified directory. By
  default, the function uses the MATCH timestamp generated by the Li-Cor system to
  identify which 'empty chamber' file is to use to correct the A-Ci measurements. If
  the Li-Cor MATCH function is not used, the option can be turned off and the closest
  empty chamber file will be used.
}
\note{
TODO: 6400 adaptation
}
\examples{

}
