% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_sample_fct.r
\name{diagnose_sample}
\alias{diagnose_sample}
\title{diagnose_sample function}
\usage{
diagnose_sample(results, sample_name, delta_max)
}
\arguments{
\item{results}{list generated by the main function Rapid_aci_correction}

\item{sample_name}{sample_ID}

\item{delta_max}{delta_max as specified to obtain results from the Rapid_aci_correction
function}
}
\value{
Several plots for diagnosing potential failure or problems
}
\description{
diagnose_sample function
}
\note{
IN CONSTRUCTION
}
