% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_aci_fct.r
\name{extr_aci}
\alias{extr_aci}
\title{extr_aci function}
\usage{
extr_aci(result_list)
}
\arguments{
\item{result_list}{The name of the resulting list produced by Rapid_aci_correction
function}
}
\value{
A list of dataframes including the corrected A and Ci, as well as other
  variables, that are used by plantecophys
}
\description{
Hidden function (for now) that could help build a data.frame of leaf
  surface area by tree using the WinSEEDLE software
}
