% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_leafArea_fct.r
\name{extr_leafArea}
\alias{extr_leafArea}
\title{extr_timestamps function}
\usage{
extr_leafArea(WinSEEDLE_filepath)
}
\arguments{
\item{winSEEDLE_filepath}{path for a winSEEDLE file (from the working directory)}
}
\value{
tree_Id list with date and leaf area
}
\description{
Hidden function (for now) that could help build a data.frame of leaf
  surface area by tree using the WinSEEDLE software
}
