% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fromExcel_fct.r
\name{get_fromExcel}
\alias{get_fromExcel}
\title{get_fromExcel function}
\usage{
get_fromExcel(filepath, leafArea_cm2 = NA, variables = c("GasEx_A",
  "GasEx_Ci", "GasEx_gtc", "GasEx_gsw", "GasEx_TleafCnd", "Meas_CO2_r",
  "Meas_Tleaf", "Meas_Tleaf2", "Meas_Qamb_in", "Const_S"),
  show.variables.names = FALSE)
}
\arguments{
\item{filepath}{Path to an Excel file}

\item{leafArea_cm2}{Numeric value of the leaf area in the chamber (default = NA, i.e.
no correction is applied)}

\item{variables}{The list of variables to retrieve from the file}

\item{show.variables.names}{Logical argument to print the entire listof variable names
to console (default = FALSE)}
}
\value{
A dataframe with the variables given in arguments
}
\description{
This function allow the extraction of selected variables from an Excel file
  produced by a Li-Cor portable photosynthesis system __LI6800__. If leaf area are
  given, the data retrieved will be adjusted accordingly.
}
