% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_system_fct.r
\name{get_system}
\alias{get_system}
\title{get_system function}
\usage{
get_system(list_files)
}
\arguments{
\item{list_files}{Inherited parameter}
}
\value{
The Li-Cor system model. This should be in cell D3 (6800) or B4 (6400).
}
\description{
Internal function to retrieve Li-Cor photosynthesis system model in Excel
  files they produced
}
