% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapid_aci_correction_fct.r
\name{Rapid_aci_correction}
\alias{Rapid_aci_correction}
\title{Rapid_aci_correction function}
\usage{
Rapid_aci_correction(list_files, delta_max = 0.05, max_degree = 3,
  priority_curve = "positive")
}
\arguments{
\item{list_files}{A dataframe as generated by the build_list function}

\item{delta_max}{Difference between two measure points under which we consider the
curve as stable. If not sure, verify with the diagnostic plots produced by the
diagnose_sample function.}

\item{max_degree}{The maximum polynomial order to consider for fitting the empty
chamber measurement curve in order to correct the Rapid A-Ci measures. This parameter
takes an integer value from 1 up to 5 (as suggested by Stinziano _et al._, 2017).
Default max_degree = 3.}

\item{priority_curve}{For the very specific case when coefficients retrieved from both
the positive and the negative portions of the empty chamber A-Ci curve, a priority
can be set. This argument can only take "positive" or "negative" (default =
"postive").}
}
\value{
The function return a list of elements by A-Ci measurement file that contains
  the sample ID, the leaf area used, the timestamp, the paths to the A-Ci and empty
  chamber file used, important variable data from these files (the recalculated ones if
  leaf area were provided), the coefficients of the fitting curve(s) used to correct
  values, as well as corrected A and Ci. The formatted data for plantecophys can be
  found under the Raci element.
}
\description{
Main function of the package. From a list of files, as produced by the
  buid_list function, the function will first extract recalculated measurement values
  when new leaf surface area is provided (or as is, if leafArea_mm2 = NA). Then the
  values will be corrected according to matching empty chamber measurements. Finally,
  data are prepared for further modelling analysis in an R list object.
}
