% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_coefs_fct.r
\name{best_coefs}
\alias{best_coefs}
\title{best_coefs function}
\usage{
best_coefs(empty_file, leafArea_cm2, delta_max, skip_first, min_CO2,
  max_degree, diagnostic_plots)
}
\arguments{
\item{empty_file}{Inherited argument from Rapid_aci_correction function}

\item{leafArea_cm2}{Leaf area in cm2}

\item{delta_max}{Inherited argument from Rapid_aci_correction function}

\item{skip_first}{Inherited argument from Rapid_aci_correction function}

\item{min_CO2}{Inherited argument from Rapid_aci_correction function}

\item{max_degree}{Inherited argument from Rapid_aci_correction function}

\item{diagnostic_plots}{Inherited argument from Rapid_aci_correction function}
}
\value{
Coefficients from the best
}
\description{
best_coefs function
}
