% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_coefs_fct.r
\name{best_coefs}
\alias{best_coefs}
\title{best_coefs function}
\usage{
best_coefs(empty_file, leafArea_cm2 = NA, delta_max, max_degree)
}
\arguments{
\item{empty_file}{Inherited argument from Rapid_aci_correction function}

\item{leafArea_cm2}{Leaf area in cm2}

\item{delta_max}{Inherited argument from Rapid_aci_correction function}

\item{max_degree}{Inherited argument from Rapid_aci_correction function}
}
\value{
Data and coefficients for the best fitting polynomial
}
\description{
Internal function to retrieve data from empty chamber measurement files
  (section 1) and coefficients for the best fitting polynomial (section 2)
}
