% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_fit_fct.r
\name{best_fit}
\alias{best_fit}
\title{best_fit function}
\usage{
best_fit(data, max_degree)
}
\arguments{
\item{data}{Data points to fit the}

\item{max_degree}{Inherited argument from Rapid_aci_correction function}
}
\value{
The list of coefficients of the best fitted curve
}
\description{
best_fit function
}
