% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_list_fct.r
\name{build_list}
\alias{build_list}
\title{build_list function}
\usage{
build_list(path_to_licor_files = "data/",
  pattern_empty = "(mpty).*\\\\.xls",
  pattern_rapidACi = "(fast).*\\\\.xls",
  pattern_standardACi = "(slow).*\\\\.xls",
  pattern_dark = "(dark).*\\\\.xls",
  sampleID_format = "[:upper:]{3}_[:digit:]{3}",
  timestamp_column = "BN", leafArea_df = NULL)
}
\arguments{
\item{path_to_licor_files}{Directory path where all files are stored}

\item{pattern_empty}{Regex pattern that must only match filenames for empty chamber
files}

\item{sampleID_format}{Regex pattern that uniquely identifies the sample ID in
filenames. For example (default), "[:upper:]{3}_[:digit:]{3}" will extract sample ID
of the format ABC_123 from a filename like
"2019-03-20_456_Logdata_ABC_123_Fast_KF.xlsx"}

\item{timestamp_column}{Column index corresponding to the MATCH TIME column in all
Li-Cor Excel files. By default this corresponds to column BN (index = 66) in Excel
files.}

\item{leafArea_df}{A dataframe containing at least a "sample_ID" column and a
"leafArea_mm2" column (default = NULL)}

\item{pattern_rapid}{Regex pattern that must only match filenames for rapid A-Ci
measurement files}

\item{pattern_slow}{Regex pattern that must only match filenames for standard A-Ci
measurement files}

\item{list_files}{A list of files as generated by the build_list function}
}
\value{
The function return a dataframe that includes the path to files, the Li-Cor
  system used, the type of measurements, the starting time of the measure, the
  timestamps, and how the timestamp was acquired.
}
\description{
Generates a list of the files present in a specified directory. By
  default, the function uses the MATCH timestamp generated by the Li-Cor system to
  identify which 'empty chamber' file is to use to correct the A-Ci measurements. If
  the Li-Cor MATCH function is not used, the option can be turned off and the closest
  empty chamber file will be used.
}
\note{
TODO: 6400 adaptation to extract START_time and timestamps
}
\examples{

}
