% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_list_fct.r
\name{build_list}
\alias{build_list}
\title{build_list function}
\usage{
build_list(path_to_licor_files = "data/",
  pattern_for_empty = "*[^(empty)]+.xlsx",
  pattern_for_measure = "*[(empty)]+.xlsx",
  match_timestamp_column = 66)
}
\arguments{
\item{path_to_licor_files}{Path where all files are stored (Empty)}

\item{pattern_for_empty}{Regex pattern that will match only filenames for empty chamber
files}

\item{pattern_for_measure}{Regex pattern that will only match filenames for measurement
files}

\item{match_timestamp_column}{Column index number that corresponds to the MATCH timestamp
in the Li-Cor measurement files. By default it should corresponds to the parameter default}
}
\description{
Generate a list of the files present in a specified directory.
The function use the MATCH timestamp to identify which 'empty chamber' file to use to
correct the measurements files (same timestamp)
}
\section{TODO}{
 find better pattern

 behavior for absence of files

 warnings for partial filename pattern match
}

