% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_dark_fct.r
\name{correct_dark}
\alias{correct_dark}
\title{correct_dark function}
\usage{
correct_dark(darkfile_path, leafArea, LiCor_system)
}
\arguments{
\item{darkfile_path}{}

\item{leafArea}{}

\item{LiCor_system}{}
}
\value{
A subset dataframe with the variables given in arguments
}
\description{
correct_dark function
}
