% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_sample_fct.r
\name{diagnose_sample}
\alias{diagnose_sample}
\title{diagnose_sample function}
\usage{
diagnose_sample(data, sample_name)
}
\arguments{
\item{data}{A list generated by the main function Rapid_aci_correction}

\item{sample_name}{sample_ID}
}
\value{
Several plots for diagnosing potential failures or problems
}
\description{
diagnose_sample function
}
\note{
IN CONSTRUCTION
}
