% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_sample_fct.r
\name{diagnose_sample}
\alias{diagnose_sample}
\title{diagnose_sample fucntion}
\usage{
diagnose_sample(results, sample_name, delta_max)
}
\arguments{
\item{results}{}
}
\description{
diagnose_sample fucntion
}
