% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_sample_fct.r
\name{diagnose_sample}
\alias{diagnose_sample}
\title{diagnose_sample function}
\usage{
diagnose_sample(data, sample_name, delta_max = 0.05,
  priority_curve = "positive")
}
\arguments{
\item{data}{list generated by the main function Rapid_aci_correction}

\item{sample_name}{sample_ID}

\item{delta_max}{delta_max as specified to obtain results from the Rapid_aci_correction
function}

\item{priority_curve}{For the very specific case when coefficients are retrieved from
both up and down ramps of the empty chamber A-Ci curve and you want to use the
coefficients from the shortest of both curve (default = "postive").}
}
\value{
Several plots for diagnosing potential failures or problems
}
\description{
diagnose_sample function
}
\note{
IN CONSTRUCTION
}
