% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_leafArea_fct.r
\name{extr_leafArea}
\alias{extr_leafArea}
\title{extr_leafArea function}
\usage{
extr_leafArea(WinSEEDLE_filepath)
}
\arguments{
\item{WinSEEDLE_filepath}{path for a winSEEDLE file (from the working directory)}
}
\value{
A dataframe of sample_ID with date and leaf area
}
\description{
Hidden function. Retrieve a dataframe of leaf surface area by sample from
  the WinSEEDLE software output files
}
