% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_timestamps_fct.r
\name{extr_timestamps}
\alias{extr_timestamps}
\title{extr_timestamps function}
\usage{
extr_timestamps(list_files, timestamp_column)
}
\arguments{
\item{match_timestamp}{}
}
\value{

}
\description{
extr_timestamps function
}
