% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_timestamps_fct.r
\name{extr_timestamps}
\alias{extr_timestamps}
\title{extr_timestamps function}
\usage{
extr_timestamps(list_files, timestamp_column)
}
\arguments{
\item{list_files}{Inherited argument from Rapid_aci_correction function}

\item{timestamp_column}{Inherited argument from Rapid_aci_correction function}
}
\value{
Timestamp as read in Excel files produced by the LI-Cor portable photosynthesis
  systems LI-6800 (only).
}
\description{
Internal function used to extract timestamps from Excel files generated by
  a Li-Cor LI-6800 system
}
