% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_timestamps_fct.r
\name{extr_timestamps}
\alias{extr_timestamps}
\title{extr_timestamps function}
\usage{
extr_timestamps(list_files, timestamp_column)
}
\arguments{
\item{list_files}{Inherited argument from Rapid_aci_correction function}

\item{timestamp_column}{Inherited argument from Rapid_aci_correction function}
}
\value{
Return file timestamp as read in Excel files produced by the LI-Cor portable
  photosynthesis systems.
}
\description{
extr_timestamps function
}
