% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapid_aci_correction_fct.r
\name{Rapid_aci_correction}
\alias{Rapid_aci_correction}
\title{Rapid_aci_correction function}
\usage{
Rapid_aci_correction(list_files, delta_max = 0.05, skip_first = 10,
  min_CO2 = 20, max_degree = 3, priority_curve = "positive",
  diagnostic_plots = FALSE)
}
\arguments{
\item{list_files}{A dataframe as generated by the build_list function}

\item{delta_max}{Difference between two measure points under which we consider the
curve as stable. If not sure, verify the diagnostic plots to see if correct}

\item{skip_first}{The number of measures to drop at the begining of measurements
(default = 10)}

\item{min_CO2}{Minimum value of CO2 that is considered correct}

\item{max_degree}{The maximum polynomial order to consider for fitting the empty
chamber measurement curve in order to correct the Rapid A-Ci measures. This parameter
takes value from 1 up to 5 (as suggested by Stinziano _et al._, 2017).}

\item{priority_curve}{For the very specific case when coefficients can be retrieve from
both the positive and the negative portion of the empty chamber A-Ci curve, a
priority has to be set. This argument can only take "positive" or "negative" (default
= "postive").}

\item{diagnostic_plots}{Output a series of plot to diagnose and verify that the
portions of the measures that are used to obtain the correction coefficients from the
empty chamber are correct}
}
\value{
The function return a list of elements by A-Ci measurement file that contains
  the sample ID, the leaf area used, the timestamp, the paths to the A-Ci and empty
  chamber file used, important data associated with the empty chamber file (the
  recalculated ones if leaf area were provided), the lag between the curves, the
  coefficients of the fitting curve used to correct values, as well as corrected Aleaf
  and Ci. correction for the Rapid A-Ci curve as well as the original values and
  parameters used for the correction.
}
\description{
Rapid_aci_correction function
}
