% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapid_aci_correction_fct.r
\name{Rapid_aci_correction}
\alias{Rapid_aci_correction}
\title{Rapid_aci_correction function}
\usage{
Rapid_aci_correction(list_files, delta_max = 0.05, skip_first = 10,
  min_CO2 = 20, max_degree = 3, diagnostic_plots = FALSE,
  curve_plot = FALSE)
}
\arguments{
\item{list_files}{A list of files as generated by the build_list function}

\item{delta_max}{Difference between two measure points under which we consider the
curve as stable. If not sure, verify the diagnostic plots to see if correct}

\item{skip_first}{The number of measures to drop at the begining of measurements (default = 10)}

\item{min_CO2}{Minimum value of CO2 that is considered correct}

\item{max_degree}{The maximum polynomial order to consider for fitting the empty
chamber measurement curve in order to correct the Rapid A-Ci measures. This parameter takes value
of 2, 3, 4 or 5 only as suggested by Stinziano _et al._ (2017).}

\item{diagnostic_plots}{Output a series of plot to diagnose/verify the process}
}
\value{
The function return a list of elements for each measurement file that contains
the corrected Rapid A-Ci curve as well as the original values and parameters used for the
correction.
}
\description{
Rapid_aci_correction function
}
