% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapid_aci_correction_fct.r
\name{Rapid_aci_correction}
\alias{Rapid_aci_correction}
\title{Rapid_aci_correction function}
\usage{
Rapid_aci_correction(list_files, delta_max = 0.05, skip_first = 10,
  min_CO2 = 20, max_degree = 3, diagnostic_plot = FALSE,
  curve_plot = FALSE)
}
\arguments{
\item{list_files}{A list of files as generated by the build_list function}

\item{delta_max}{A difference between two measure points under which we consider the
curve as stable. If not sure, verify the diagnostic_plot to see if correct}

\item{skip_first}{The number of measures to drop at the beginning of measurements}

\item{min_CO2}{Minimum value of CO2 that is considered real}

\item{max_degree}{THe maximum order of polynomial to consider for fitting the empty
chamber curve use as correction for Rapid A-Ci measurements. This can take a value between 2 and 5.}

\item{diagnostic_plot}{Output a series of plot to diagnose/verify the process}

\item{curve_plot}{Will be removed eventually...}
}
\description{
Rapid_aci_correction function
}
